Major testing changes, not yet working again fully
Added handling for endbr32 instructions.

Testing should happen closer on/to the 'live' executable (eg. in terms of
compilation flags only used on the infect target, but not the infect
code compiled for the testing binaries).
For this, a raw binary of the infect code with all used compilation
flags is create and introduced for usage in the tests.

The Function discovery test is yet the only one which is adapted so far.
It now also really checks for all functions appearing in the code.